refactor(daemon-config): move PipeConfig and NotificationConfig to more specific folders in the configuration
