+++ src/daemon/Tiempitod.NET/Program.cs
@@ -8,6 +8,8 @@ using Tiempitod.NET.Commands.Handler;
 using Tiempitod.NET.Commands.Server;
 using Tiempitod.NET.Configuration;
 using Tiempitod.NET.Configuration.AppFilesystem;
+using Tiempitod.NET.Configuration.Notifications;
+using Tiempitod.NET.Configuration.Server;
 using Tiempitod.NET.Configuration.Session;
 using Tiempitod.NET.Configuration.User;
 using Tiempitod.NET.Notifications;
